IModel refactoring

+++ hummer-core/src/test/java/org/hummer/newweb/dao/user/impl/TestDAO.java
@@ -16,7 +16,7 @@ public class TestDAO extends BasicMybatisDAO implements ITestDAO {
     }
 
     @Override
-    public IUser getUserById(Integer id) {
+    public IUser getUserById(Long id) {
         return (IUser) super.getModel(id);
     }
