+++ hummer-core/src/test/java/org/hummer/newweb/dao/user/intf/ITestDAO.java
@@ -6,5 +6,5 @@ import org.hummer.newweb.model.intf.IUser;
 public interface ITestDAO extends IBasicTestDAO {
     int insert(IUser a);
 
-    IUser getUserById(Integer id);
+    IUser getUserById(Long id);
 }
